Add protected-headers="v1" to Content-Type when protecting headers
This is optional part of the draft RFC, marked with "FIXME":

    * FIXME: Enigmail adds "protected-headers="v1"" parameter to
      "payload" here. Is this necessary?

The answer is: for Enigmail yes. Otherwise it ignores protected headers
and use the envelope headers. This results in all the emails listed with
"..." subject, which isn't very helpful.

Since the user can disable protected headers while in the compose menu
(after potentially failed send), remove the attribute if protected
headers are disabled.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/162e60883c76071641dcc46ef09bd596f50068a3
Co-authored-by: Richard Russon <rich@flatcap.org>