Fix crash when polling a closed ssl connection
Commit 8353407c enabled checking for buffered OpenSSL/GnuTLS data when
polling, but neglected to check if the connection was already closed.

This can be triggered during imap_logout() if the connection write of
"LOGOUT" fails and closes the connection, before the poll.  It's a bit
tricky to trigger because imap_logout_all() checks for a closed
connection, so the failure needs to take place at that last write.

Thanks to Stefan Sperling for pointing out the problem, complete with
a backtrace and patch. (This commit takes a different approach for a
stable-branch fix.)

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/edf5699c189bf8da642297fe327e19a6c7674091
Co-authored-by: Richard Russon <rich@flatcap.org>